Start the dependent applications, not they supervisors.
Instead of adding the partisan and plumtree supervisor tree to the
plum_db supervisor tree, start these applications on their own. This
makes it possible to use the plum_db in other applications depending on
partisan (that other application starts partisan on its own, so starting
its supervisor would lead to name clash and crash).

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>

# Conflicts:
#	src/plum_db_app.erl