latency reduction by removing worker and allowing direct access to databases on gets
We removed the partition worker process which was there to serialise read-modify-write logic. It was unnecesary as moving all the logic to partition_server reduces latency dramatically.

Calls to get now accepts options. If option `allow_put` has value `false` then the read (and any resolving) is performed by the calling process reducing latency even further.  A new shared table containing the new db_info record was introduce to support this feature. If `allow_put` has value `true` then a call is made to the partition server. This is to avoid concurrent calls resolving conflicts to produce more conflicts. More analysis needs to be done to find additional safe options for this particular case.

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>