fixes non-deterministic binary serialization of Erlang terms, introduced in latest OTP version
Running the followin on two different nodes can yiled different results even when the Object is identical. hash() uses a crypto sha of the binary representation of O.

plum_db_object:hash(plum_db:get_object(PrefixedKey}))

The cause is that term_to_binary can yield different binaruies on different nodes (maybe due to the object containing maps for which the internal order is undefined and proven in this case or something else affecting the serialization).

The solution is to pass the `deterministic` option to `term_to_binary`

Signed-off-by: Alejandro M. Ramallo <alejandro.ramallo@leapsight.com>